refactored tests and removed comments
