+++ libs/designsystem/src/lib/components/text-link/text-link.component.spec.ts
@@ -8,24 +8,8 @@ import { SizeDirective } from '../../directives';
 import { IconComponent } from '../icon';
 
 const getColor = DesignTokenHelper.getColor;
-const fontSize = DesignTokenHelper.fontSize;
-const lineHeight = DesignTokenHelper.lineHeight;
-const size = DesignTokenHelper.size;
 
-/*
-TEST TO INCLUDE
-1.  Test for size - xs,sm,md for both -> icon √ and text √
-2.  Test for Icon apperance on external links √
-3.  Test logic for determining external links?
-4.  Test for right color (1c1c1c) √
-5.  Test for text containg  underline √
-6.  Test for using the correct icon √
-7.  Test that the cursor changes when mouse over link? √
-8.  Test that Icon is position to the right from text √
-9.  Test that text-link component is created  √
-
-*/
-describe('TextLinkComponent', () => {
+describe('TextLinkComponent unitTest', () => {
   let spectator: SpectatorHost<TextLinkComponent>;
 
   const createHost = createHostFactory({
@@ -57,39 +41,8 @@ describe('TextLinkComponent', () => {
     const link = spectator.hostElement.getElementsByTagName('kirby-text-link');
     expect(link[0]).toHaveComputedStyle({ cursor: 'pointer' });
   });
-  describe('Test TextLink sizes', () => {
-    interface TestCase {
-      name: string;
-      value: string;
-      expected: string;
-    }
-    const testCases: TestCase[] = [
-      { name: 'md', value: 'md', expected: 'n' },
-      { name: 'sm', value: 'sm', expected: 's' },
-      { name: 'xs', value: 'xs', expected: 'xs' },
-
-      { name: 'Empty', value: '', expected: 'n' },
-    ];
-    testCases.forEach((tc) => {
-      it('should render correct size & line-height, when size is ' + tc.name, () => {
-        spectator = createHost(`<kirby-text-link size='${tc.value}' text='Some Link' link='/'>`);
-        const link = spectator.queryHost<HTMLElement>('a');
-        expect(link).toHaveComputedStyle({
-          'font-size': fontSize(tc.expected),
-          'line-height': lineHeight(tc.expected),
-        });
-      });
-    });
-  });
 
   describe('Test TextLink with icon', () => {
-    interface TestCase {
-      name: string;
-      value: string;
-      expected_font: string;
-      expected_padding_right: string;
-      expected_padding_left: string;
-    }
     it('should render a text link with a icon', () => {
       spectator = createHost(
         `<kirby-text-link  text='Some Link' link='https://angular.io/api/router/RouterLink'></kirby-text-link>`
@@ -116,49 +69,6 @@ describe('TextLinkComponent', () => {
       const textPosition = text.getBoundingClientRect();
       const iconPosition = icon.getBoundingClientRect();
       expect(textPosition.right).toBeLessThan(iconPosition.right);
-    });
-    const testCases: TestCase[] = [
-      {
-        name: 'md',
-        value: 'md',
-        expected_font: 'm',
-        expected_padding_left: 'xxxs',
-        expected_padding_right: 's',
-      },
-      {
-        name: 'sm',
-        value: 'sm',
-        expected_font: 'm',
-        expected_padding_left: 'xxxs',
-        expected_padding_right: 's',
-      },
-      {
-        name: 'xs',
-        value: 'xs',
-        expected_font: 's',
-        expected_padding_left: 'xxxxs',
-        expected_padding_right: 's',
-      },
-      {
-        name: 'Empty',
-        value: '',
-        expected_font: 'm',
-        expected_padding_left: 'xxxs',
-        expected_padding_right: 's',
-      },
-    ];
-    testCases.forEach((tc) => {
-      it('should render correct size & padding, when size is ' + tc.name, () => {
-        spectator = createHost(
-          `<kirby-text-link size='${tc.value}' text='Some Link' link='https://angular.io/api/router/RouterLink'></kirby-text-link>`
-        );
-        const icon = spectator.queryHost<HTMLElement>('kirby-icon');
-        expect(icon).toHaveComputedStyle({
-          'font-size': size(tc.expected_font),
-          'padding-left': size(tc.expected_padding_left),
-          'padding-right': size(tc.expected_padding_right),
-        });
-      });
     });
   });
 });
